% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diseñocuadradolatino.R
\name{ADCL}
\alias{ADCL}
\title{Tabla ANOVA de un Diseño Cuadrado Latino}
\usage{
ADCL(respuesta, tratamiento, columna, fila, data)
}
\arguments{
\item{respuesta}{(string) nombre de la variable respuesta.}

\item{tratamiento}{(string) nombre de la variable que representan a los tratamientos.}

\item{columna}{(string) nombre de la variable que representa a las columnas.}

\item{fila}{(string) nombre de la variable que representa a las filas.}

\item{data}{(\code{data.frame}) Tabla de los datos en formato largo con los datos de los tratamientos, columnas, filas y de la variable respuesta.}
}
\value{
Devuelve una tabla en formato \code{data.frame} con los cálculos correspondientes al análisis de varianza de un Diseño Cuadrado Latino.
}
\description{
Genera la tabla de Análisis de Varianza (ANOVA) para un Diseño Cuadrado Latino (DCL)
}
\examples{
\dontrun{
## 1 Directorio de trabajo
ruta<-"D:/ruta/test 1.csv"
df<-read.csv(ruta)

## 2 Cargo la librería
library(DCL)

## Ejecutamos la función ANOVA_DCL
ADCL(respuesta = "Respuesta",tratamiento = "Tratamiento",columna = "Columna",fila = "Fila",data = df)
}
}
